feat: replace time block form with modal
