+++ src/app/components/dashboard-component/dashboard-component.ts
@@ -22,7 +22,7 @@ import { ServicesService } from '../../services/services-service';
 import { TimeBlock, TimeBlockService } from '../../services/time-block-service';
 import { ToastService } from '../../services/toast-service';
 import { AppointmentFormComponent } from '../../components/appointment-form-component/appointment-form-component';
-import { TimeBlockFormComponent } from '../../components/time-block-form-component/time-block-form-component';
+import { BlockTimeModalComponent } from '../../components/block-time-modal/block-time-modal';
 import { AppointmentDetailsModalComponent } from '../../components/appointment-details-modal-component/appointment-details-modal-component';
 
 @Component({
@@ -32,7 +32,7 @@ import { AppointmentDetailsModalComponent } from '../../components/appointment-d
     CommonModule,
     RouterModule,
     AppointmentFormComponent,
-    TimeBlockFormComponent,
+    BlockTimeModalComponent,
     AppointmentDetailsModalComponent,
   ],
   templateUrl: './dashboard-component.html',
